% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinybootstrap2.R
\docType{package}
\name{shinybootstrap2-package}
\alias{shinybootstrap2}
\alias{shinybootstrap2-package}
\title{Bootstrap 2 components for use with Shiny}
\description{
Bootstrap 2 components for use with Shiny
}

