% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input-text.R
\name{textInput}
\alias{textInput}
\title{Create a text input control}
\usage{
textInput(inputId, label, value = "")
}
\arguments{
\item{inputId}{Input variable to assign the control's value to}

\item{label}{Display label for the control}

\item{value}{Initial value}
}
\value{
A text input control that can be added to a UI definition.
}
\description{
Create an input control for entry of unstructured text values
}
\examples{
textInput("caption", "Caption:", "Data Summary")
}
\seealso{
\code{\link[shiny]{updateTextInput}}

Other input.elements: \code{\link{animationOptions}},
  \code{\link{sliderInput}};
  \code{\link{checkboxGroupInput}};
  \code{\link{checkboxInput}}; \code{\link{dateInput}};
  \code{\link{dateRangeInput}}; \code{\link{numericInput}};
  \code{\link{radioButtons}}; \code{\link{selectInput}},
  \code{\link{selectizeInput}}; \code{\link{submitButton}}
}

