% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/output-datatable.R
\name{dataTableOutput}
\alias{dataTableOutput}
\title{Create a table output element}
\usage{
dataTableOutput(outputId)
}
\arguments{
\item{outputId}{output variable to read the table from}
}
\value{
A table output element that can be included in a panel
}
\description{
Render a \code{\link{renderDataTable}} within an application page.
\code{renderDataTable} uses the DataTables Javascript library to create an
interactive table with more features.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  library(shiny)
  shinyApp(
    ui = bootstrapPage(
      dataTableOutput('table')
    ),
    server = function(input, output) {
      output$table <- renderDataTable(iris)
    }
  )
}
}

